move DataManagementContent to DataManagementContent folder
