+++ workbase/src/renderer/routes.js
@@ -3,7 +3,7 @@ import MainTemplate from './components/MainTemplate/MainTemplate.vue';
 import SchemaDesignContent from './components/SchemaDesign/SchemaDesignContent/SchemaDesignContent.vue';
 import SettingsPage from './components/ManageSettings/SettingsPage.vue';
 import KeyspacesPage from './components/ManageKeyspaces/KeyspacesPage.vue';
-import DataManagementPage from './components/DataManagement/DataManagementContent.vue';
+import DataManagementPage from './components/DataManagement/DataManagementContent/DataManagementContent.vue';
 import LoginPage from './components/Login/LoginPage.vue';
 
 export const isDataManagement = route => (route.path === '/develop/data');
